+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/library/bookmarks/BookmarkDeselectNavigationListenerTest.kt
@@ -28,6 +28,7 @@ class BookmarkDeselectNavigationListenerTest {
         title = null,
         url = null,
         dateAdded = 0,
+        lastModified = 0,
         children = null,
     )