+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/library/bookmarks/BookmarkFragmentInteractorTest.kt
@@ -32,9 +32,9 @@ class BookmarkFragmentInteractorTest {
 
     private val bookmarkController: DefaultBookmarkController = mockk(relaxed = true)
 
-    private val item = BookmarkNode(BookmarkNodeType.ITEM, "456", "123", 0u, "Mozilla", "http://mozilla.org", 0, null)
-    private val separator = BookmarkNode(BookmarkNodeType.SEPARATOR, "789", "123", 1u, null, null, 0, null)
-    private val subfolder = BookmarkNode(BookmarkNodeType.FOLDER, "987", "123", 0u, "Subfolder", null, 0, listOf())
+    private val item = BookmarkNode(BookmarkNodeType.ITEM, "456", "123", 0u, "Mozilla", "http://mozilla.org", 0, 0, null)
+    private val separator = BookmarkNode(BookmarkNodeType.SEPARATOR, "789", "123", 1u, null, null, 0, 0, null)
+    private val subfolder = BookmarkNode(BookmarkNodeType.FOLDER, "987", "123", 0u, "Subfolder", null, 0, 0, listOf())
     private val tree: BookmarkNode = BookmarkNode(
         BookmarkNodeType.FOLDER,
         "123",
@@ -42,6 +42,7 @@ class BookmarkFragmentInteractorTest {
         0u,
         "Mobile",
         null,
+        0,
         0,
         listOf(item, separator, item, subfolder),
     )