+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/library/bookmarks/NavBarUtilsKtTest.kt
@@ -90,6 +90,7 @@ internal fun testBookmarkItem(parentGuid: String, url: String, title: String = "
     title,
     url,
     0,
+    0,
     null,
 )
 
@@ -101,6 +102,7 @@ internal fun testFolder(guid: String, parentGuid: String?, children: List<Bookma
     title,
     null,
     0,
+    0,
     children,
 )
 
@@ -111,6 +113,7 @@ internal fun testSeparator(parentGuid: String) = BookmarkNode(
     null,
     null,
     null,
+    0,
     0,
     null,
 )